dwc3: gadget: Handle dequeuing of non queued request gracefully
Trying to dequeue an request that is currently not queued should be a no-op
and be handled gracefully.

Checking on list/queue empty indicate whether the request is queue or not.
Handling this gracefully allows for race condition free synchronization
between the complete callback being called to to a completed transfer and
trying to call usb_ep_dequeue() at the same time.

Inspired by:
https://patchwork.kernel.org/project/linux-usb/patch/20191106144553.16956-1-alexandru.ardelean@analog.com/

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>